beginner/submissions/team-members/manan-jariwala/requirements.txt
gradio
openai
tavily-python
python-dotenv
